feat(inputValidation): 사용자 입력에서 숫자로만 이루어졌는지 검증하는 기능 추가
- 기존 ValidationUtils를 대체하여 InputValidation을 추가

기존의 ValidationUtils 이라는 이름은 무엇을 검증하는지 충분히 설명할 수 없었다. 중복이나 길이에 대한 검증은 이미 Balls나 Ball에서 담당하므로 숫자로만 이루어졌는지만 검증하는 클래스를 추가함.
